[DS-6403][fix] Put DB queries ahead before submit to TaskPriorityQueue
